DB as a standalone pkg (#8)
Prisma has nice DX for generating migrations but pain to use it in
different pkgs (need some effort) & the other big limitation is I cannot
add custom data types.

I want to use the SQlite for storing the embeddings & drizzle with turso
has better support
https://docs.turso.tech/sdk/ts/orm/drizzle#vector-embedding

So this PR splits the database package which the crawler service can
use. Later it will be easier for agent to direclty use it instead of me
creating a custom wrapper for sqlite (that i did for prisma to do it
quickly).

- [X] tested locally & works as it usually does
- [x] Need to deploy on cluster and test it 

Updated the build process for image. Now everything is bundled & native
deps are added explicitly to docker image. Makes the images smaller & in
theory could deploy easily to lambdas too now if needed :)