feat: ingesting pdf based reports (#54)
The goal with this one is to be able to have a pipeline that when given
a PDF URL be able to
1. extract all the contents
2. Sanitize it
3. Chunk it
4. Embed it

This way as soon as we ingest any new reports they are just there. This
is more raw form type of parser which I will make with the discord
management server so I can just tell bot to ingest. Eventually want to
feed it more data types but one step at a time.

- [x] `Document` table -- this will be used to store the document
information and any metadata necessary for it
- [ ] Rename `BillVector` table to `Vector` 
- [x] Alter `Vector` to have relation with `Document` table for future
joins
- [x] Run prod migrations
- [x] `crawler` deploy
- [x] `agent` deploy


closes #38