feat: integrate web scraper results (#99)
This builds on top of #98 and uses the embeddings we stored for web page
data and inject that as context for LLMs to use to generate responses.


Refactor the code to have a centralize querying for the knowledge base.
Need to refactor further so each type of interacts we support is a
module of it's own to avoid duplication.