feat: only crawl bill if it changed in last week (#241)
a bill text can update or bill action can update. So better is to look
at when it was last updated then only enqueue to process it. This way we
can make sure we have up to date and only include the bills that make
sense to be included.

A week (8 days as buffer) make sense since we run this every week.